requirements.txt
folium==0.17.0
haversine==2.8.1
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
plotly==5.24.1
streamlit==1.38.0
streamlit_folium==0.22.1
